refactor nav menu into a reusable component
