+++ code/src/ui/pages/landing.tsx
@@ -7,18 +7,11 @@ import "github-markdown-css/github-markdown.css";
 import styles from "./landing.module.css";
 
 import markdownit from "markdown-it";
-import { useMemo } from "react";
 
-import { usePermissionScheme } from "hooks/auth/usePermissionScheme";
-import { Assets } from "io/config/assets";
 import { OptionalPage } from "io/config/optional-pages";
-import { Modules, Routes } from "io/config/routes";
-import { PermissionScheme } from "types/auth";
 import { Dictionary } from "types/dictionary";
-import { NavBarItemSettings, UISettings } from "types/settings";
-import { NavBarUploadItem } from "ui/navigation/navbar/navbar-upload-item";
-import { parseStringsForUrls, parseWordsForLabels } from "utils/client-utils";
-import { NavBarItem } from "../navigation/navbar/navbar-item";
+import { UISettings } from "types/settings";
+import { NavMenu } from "ui/navigation/navbar/nav-menu";
 
 // Utilities to render markdown into HTML
 const markdowner = markdownit({
@@ -42,146 +35,18 @@ interface LandingPageProps {
  * @returns JSX for landing page.
  */
 export default function LandingPage(props: Readonly<LandingPageProps>) {
-  const ASSET_PREFIX = process.env.ASSET_PREFIX ?? "";
-  const keycloakEnabled = process.env.KEYCLOAK === "true";
   // CSS class names
   const introClasses = ["markdown-body", styles.introInner].join(" ");
-  const permissionScheme: PermissionScheme = usePermissionScheme();
-  // Retrieve links
-  const dashboardLinkProps: NavBarItemSettings = props.settings.links?.find(
-    (link) => link.url === Modules.DASHBOARD
-  );
-  const helpLinkProps: NavBarItemSettings = props.settings.links?.find(
-    (link) => link.url === Modules.HELP
-  );
-  const mapLinkProps: NavBarItemSettings = props.settings.links?.find(
-    (link) => link.url === Modules.MAP
-  );
-  const registryLinkProps: NavBarItemSettings = props.settings.links?.find(
-    (link) => link.url === Modules.REGISTRY
-  );
-  const registryUrl: string = useMemo(() => {
-    // Defaults to pending registry with no route or scheme is disabled
-    let url: string = `${Routes.REGISTRY_PENDING}/${props.settings.resources?.registry?.data}`;
-    if (permissionScheme?.registryPageLink) {
-      url = permissionScheme?.registryPageLink;
-      // Only update the permission route if they are pending or active
-      if (url === Routes.REGISTRY_PENDING || url === Routes.REGISTRY_ACTIVE) {
-        url = `${url}/${props.settings.resources?.registry?.data}`;
-      }
-    }
-    return url;
-  }, [permissionScheme]);
 
   return (
     <div className="flex h-screen w-full">
       {/* This is the navigation on the left */}
-      <div className="bg-muted  border-r-border hidden w-3xs flex-col items-center justify-start gap-6 overflow-x-scroll overflow-y-auto border-r pb-20 lg:w-xs xl:flex 2xl:w-xs">
-        {props.pages
-          .filter((page) => page.slug !== "landing" && page.slug !== "help")
-          .map((page) => (
-            <NavBarItem
-              key={page.title}
-              title={page.title}
-              icon={page.thumbnail ?? Assets.INFO}
-              url={`${ASSET_PREFIX}/${page.slug}`}
-              isMobile={false}
-              caption={page.description}
-            />
-          ))}
-
-        {props.settings.modules.map && (
-          <NavBarItem
-            title={mapLinkProps?.title ?? props.dict.nav.title.map}
-            icon={mapLinkProps?.icon ?? Assets.MAP}
-            url={Routes.MAP}
-            isMobile={false}
-            caption={mapLinkProps?.caption ?? props.dict.nav.caption.map}
-          />
-        )}
-        {props.settings.modules.dashboard && (
-          <NavBarItem
-            title={dashboardLinkProps?.title ?? props.dict.nav.title.dashboard}
-            icon={dashboardLinkProps?.icon ?? Assets.DASHBOARD}
-            url={Routes.DASHBOARD}
-            isMobile={false}
-            caption={
-              dashboardLinkProps?.caption ?? props.dict.nav.caption.dashboard
-            }
-          />
-        )}
-        {props.settings.modules.registry &&
-          props.settings.resources?.registry?.data && (
-            <NavBarItem
-              title={registryLinkProps?.title ?? props.dict.nav.title.registry}
-              icon={registryLinkProps?.icon ?? Assets.REGISTRY}
-              url={registryUrl}
-              isMobile={false}
-              caption={
-                registryLinkProps?.caption ?? props.dict.nav.caption.registry
-              }
-            />
-          )}
-        {props.settings.modules.registry &&
-          props.settings.resources?.registry?.paths?.map((path, index) => (
-            <NavBarItem
-              key={path + index}
-              title={parseWordsForLabels(path)}
-              icon={Assets.REGISTRY}
-              url={`${Routes.REGISTRY_GENERAL}/${parseStringsForUrls(path)}`}
-              isMobile={false}
-              caption={props.dict.nav.caption.generalReg.replace(
-                "{replace}",
-                parseWordsForLabels(path).toLowerCase()
-              )}
-            />
-          ))}
-
-        {props.settings.modules.help && (
-          <NavBarItem
-            title={helpLinkProps?.title ?? props.dict.nav.title.help}
-            icon={helpLinkProps?.icon ?? Assets.HELP}
-            url={Routes.HELP}
-            isMobile={false}
-            caption={helpLinkProps?.caption ?? props.dict.nav.caption.help}
-          />
-        )}
-
-        {props.settings.links?.map((externalLink, index) => {
-          if (
-            ![
-              Modules.MAP,
-              Modules.DASHBOARD,
-              Modules.HELP,
-              Modules.REGISTRY,
-            ].includes(externalLink.url) &&
-            // When authentication is disabled OR no permission is set for this button in the UI-Settings, all users can view and access these buttons
-            // IF there is a permission set with authentication enabled, check if the user has the specified permission
-            (!keycloakEnabled ||
-              !externalLink?.permission ||
-              permissionScheme?.hasPermissions[externalLink.permission])
-          ) {
-            return (
-              externalLink.type === "file" ? <NavBarUploadItem
-                key={externalLink.title + index}
-                title={externalLink.title}
-                icon={externalLink.icon}
-                url={externalLink.url}
-                isMobile={false}
-                caption={externalLink.caption}
-              /> : <NavBarItem
-                key={externalLink.title + index}
-                title={externalLink.title}
-                icon={externalLink.icon}
-                url={externalLink.url}
-                isMobile={false}
-                caption={externalLink.caption}
-              />
-            );
-          }
-        })}
-      </div>
-      {/* This is the where the tumnnail and Project outline is */}
+      <NavMenu
+        settings={props.settings}
+        isMobile={false}
+        pages={props.pages}
+      />
+      {/* This is the where the thumbnail and Project outline is */}
       <div className="mx-auto mt-4 flex h-4/5 items-center justify-center p-2 sm:w-sm md:h-11/12 md:w-11/12 lg:h-11/12 lg:w-11/12 lg:p-4 xl:mt-0 xl:h-10/12 xl:w-9/12 2xl:h-11/12">
         <div className="bg-muted border-border flex h-full flex-col rounded-xl border-1 p-5 shadow-2xl">
           <div
